Gui merge

+++ GradingProject/src/projectfiles/GradingCategory.java
@@ -1,16 +1,15 @@
 package projectfiles;
 
-import java.awt.Component;
 import java.util.LinkedList;
 
 public class GradingCategory {
 	//instance variables
 	public String name;
-	public int percentage;
+	public double percentage;
 	public LinkedList<Assignment> assignments;
 	
 	//constructor
-	public GradingCategory(String name, int percentage) {
+	public GradingCategory(String name, double percentage) {
 		this.name = name;
 		this.percentage = percentage;
 		this.assignments = new LinkedList<Assignment>();
